# build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Customer": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"company": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"company"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CustomersResponse": {
				"properties": {
					"customersList": {
						"items": {
							"$ref": "#/components/schemas/Customer"
						},
						"type": "array"
					},
					"companiesList": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"customersList",
					"companiesList"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "customers-service",
		"version": "1.0.0",
		"description": "customers-service",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/customers/fetch": {
			"get": {
				"operationId": "GetCustomers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomersResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "max",
						"required": false,
						"schema": {
							"default": 50,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"default": "%",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"default": "firstName",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "companyFilter",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}